backend/requirements.txt
asgiref==3.8.1
django==5.1.2
django-cors-headers==4.6.0
django-vite-react==0.0.3
djangorestframework==3.15.2
django-webpack-loader==3.1.1
pip==24.3.1
psycopg2-binary==2.9.10
setuptools==59.6.0
sqlparse==0.5.1
typing-extensions==4.12.2
wheel==0.45.1